docs/tools/code-index/requirements.txt
tree-sitter==0.25.2
tree-sitter-language-pack
pyyaml
